✨ feat: Pong 컴포넌트 분리 및 소켓 연결
